+++ config/di.php
@@ -2,10 +2,12 @@
 
 declare(strict_types=1);
 
+use PHPForge\Debug\Storage\SnapshotStore;
 use Psr\Http\Message\StreamFactoryInterface;
+use Yii3\Debug\ConfigDataFactory;
 use Yii3\Debug\Middleware\ToolbarMiddleware;
 use Yii3\Debug\ToolbarDataFactory;
-use Yii3\Debug\Web\ToolbarRenderer;
+use Yii3\Debug\Web\{DebugPageRenderer, ToolbarRenderer};
 use Yiisoft\Aliases\Aliases;
 use Yiisoft\Assets\AssetManager;
 use Yiisoft\NetworkUtilities\IpRanges;
@@ -15,24 +17,51 @@ use Yiisoft\View\WebView;
 $config = $params['yii3/debug'];
 
 return [
-    ToolbarDataFactory::class => [
+    ConfigDataFactory::class => [
         '__construct()' => [
-            'position' => $config['toolbar']['position'],
-            'height' => $config['toolbar']['height'],
+            'application' => $config['application'],
         ],
     ],
+    DebugPageRenderer::class => static fn(
+        WebView $view,
+        AssetManager $assetManager,
+        ConfigDataFactory $configDataFactory,
+        Aliases $aliases,
+    ): DebugPageRenderer => (
+        new DebugPageRenderer(
+            $view,
+            $assetManager,
+            $configDataFactory,
+            $aliases->get('@vendor/php-forge/debug-core/resources/views'),
+        )
+    )
+    ->withRoutePrefix($config['routePrefix']),
+    SnapshotStore::class => static fn(Aliases $aliases): SnapshotStore => new SnapshotStore(
+        path: $aliases->get($config['storage']['path']),
+        dirMode: $config['storage']['dirMode'],
+        fileMode: $config['storage']['fileMode'],
+    ),
+    ToolbarDataFactory::class => static fn(AssetManager $assetManager): ToolbarDataFactory => (
+        new ToolbarDataFactory($assetManager)
+    )
+    ->withRoutePrefix($config['routePrefix'])
+    ->withPresentation($config['toolbar']['position'], $config['toolbar']['height']),
     ToolbarMiddleware::class => static fn(
         ToolbarRenderer $renderer,
         StreamFactoryInterface $streamFactory,
-    ): ToolbarMiddleware => new ToolbarMiddleware(
-        renderer: $renderer,
-        streamFactory: $streamFactory,
-        allowedIpRanges: new IpRanges($config['allowedIPs']),
-        routePrefix: $config['routePrefix'],
-        skipUrls: $config['toolbar']['skipUrls'],
-        position: $config['toolbar']['position'],
-        height: $config['toolbar']['height'],
-    ),
+        SnapshotStore $store,
+    ): ToolbarMiddleware => (
+        new ToolbarMiddleware(
+            $renderer,
+            $streamFactory,
+            $store,
+            new IpRanges($config['allowedIPs']),
+        )
+    )
+    ->withRoutePrefix($config['routePrefix'])
+    ->withHistorySize($config['historySize'])
+    ->withSkipUrls($config['toolbar']['skipUrls'])
+    ->withPresentation($config['toolbar']['position'], $config['toolbar']['height']),
     ToolbarRenderer::class => static fn(
         WebView $view,
         AssetManager $assetManager,